reconstruction/sfm.py
import cv2
import numpy as np
from pathlib import Path
from utils.config import config
from utils.logger import logger

class SFMReconstructor:
    def __init__(self):
        self.sift = cv2.SIFT_create()
        self.matcher = cv2.FlannBasedMatcher_create()
        self.camera_matrix = np.array([
            [500, 0, 320],
            [0, 500, 240],
            [0, 0, 1]
        ])

    def process_sequence(self, image_dir, output_dir):
        images = sorted(Path(image_dir).glob("*.jpg"))
        features = []
        
        for img_path in images:
            img = cv2.imread(str(img_path))
            gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
            kp, desc = self.sift.detectAndCompute(gray, None)
            features.append((img_path, kp, desc))
        
        # Feature matching and pose estimation
        point_cloud = self._reconstruct(features)
        self._save_results(point_cloud, output_dir)

    def _reconstruct(self, features):
        # Implement incremental SfM pipeline
        pass