requirements.txt
torch 
torchvision 
torchaudio
tensorflow 
keras
numpy 
matplotlib 
opencv-python 
pillow 
tqdm
torchmetrics
timm 
open3d
scipy
jupyter
seaborn
scikit-learn
scikit-image
tinycudann